+++ src/constants.ts
@@ -1,9 +1,9 @@
-export const $signal = Symbol.for('signal');
-export const $effect = Symbol.for('effect');
+export const $signal = Symbol.for('@spearwolf/signalize/signal');
+export const $effect = Symbol.for('@spearwolf/signalize/effect');
 
-export const $destroySignal = Symbol.for('destroySignal');
-export const $createEffect = Symbol.for('createEffect');
-export const $destroyEffect = Symbol.for('destroyEffect');
+export const $destroySignal = Symbol.for('@spearwolf/signalize/destroySignal');
+export const $createEffect = Symbol.for('@spearwolf/signalize/createEffect');
+export const $destroyEffect = Symbol.for('@spearwolf/signalize/destroyEffect');
 
 export const VALUE = 'value';
 export const MUTE = 'mute';
@@ -11,4 +11,4 @@ export const UNMUTE = 'unmute';
 export const DESTROY = 'destroy';
 export const OFF = 'off';
 
-export const RECALL = Symbol.for('recall');
+export const RECALL = Symbol.for('@spearwolf/signalize/recall');